services/SafeZoneDashboard/requirements.txt
python-dotenv==1.0.1
dash==2.18.2
pandas==2.2.3
requests==2.32.3
pydantic==2.10.4
dash_bootstrap_components==1.6.0
python-json-logger==3.2.1